Split /ldsg into its own file
I just noticed that we have other subcommands that need to be added to it.
